Sound Effect Implementation
Using the animation curve of the magic and picking up animation, I added a sound effect to each at a certain point on their curves that mimic the models movements nicely.  I also added a scary sound which is programmed to occur when the enemy has the player in it's field of vision (to warn player its following them).